# Rating API/api/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API to manage ratings.",
        "title": "Rating API",
        "contact": {
            "name": "Vincent GIMBERT.",
            "email": "vincent.gimbert@yahoo.fr"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/ratings": {
            "get": {
                "description": "Get ratings.",
                "tags": [
                    "ratings"
                ],
                "summary": "Get ratings.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Rating"
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            },
            "post": {
                "description": "Create a rating. UUID is automatically generated",
                "tags": [
                    "ratings"
                ],
                "summary": "Create a rating.",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Rating"
                        }
                    },
                    "422": {
                        "description": "Cannot parse id"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            },
            "delete": {
                "description": "Delete a rating.",
                "tags": [
                    "ratings"
                ],
                "summary": "Delete a rating.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection UUID formatted ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "422": {
                        "description": "Cannot parse id"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/ratings/{id}": {
            "get": {
                "description": "Get a rating.",
                "tags": [
                    "ratings"
                ],
                "summary": "Get a rating.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection UUID formatted ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Rating"
                        }
                    },
                    "422": {
                        "description": "Cannot parse id"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            },
            "put": {
                "description": "Update a rating.",
                "tags": [
                    "ratings"
                ],
                "summary": "Update a rating.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection UUID formatted ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Cannot parse id"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Rating": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "description": "id de l'utilisateur",
                    "type": "string"
                },
                "note": {
                    "type": "integer"
                }
            }
        }
    }
}